% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pasilla}
\alias{pasilla}
\title{pasilla}
\format{
A SummarizedExperiment object with the 7 pasilla samples. Generated from the code in data-raw/pasilla.R.
}
\source{
\url{https://bioconductor.org/packages/release/data/experiment/html/pasilla.html}
}
\usage{
data(pasilla)
}
\description{
pasilla gene count data in SummarizedExperiment format
}
\keyword{datasets}
